move logic from `LanguageProvider` into `GraphQLClientProject`, delete `LanguageProvider`
